Extract wearable detail modal
