+++ tests/test-ui-flows.js
@@ -93,14 +93,14 @@ return (async function() {
   const viewsSrc = await fetch('js/views.js').then(r => r.text());
   const markerDetailSrc = await fetch('js/marker-detail-modal.js').then(r => r.text());
   const dashboardWidgetsSrc = await fetch('js/dashboard-widgets.js').then(r => r.text());
-  const wearablesSrc = await fetch('js/wearables.js').then(r => r.text());
+  const wearablesDetailSrc = await fetch('js/wearables-detail-modal.js').then(r => r.text());
   assert('marker-detail-modal.js defines rememberModalTrigger', /function rememberModalTrigger\s*\(/.test(markerDetailSrc));
   assert('marker-detail-modal.js defines restoreModalTrigger', /function restoreModalTrigger\s*\(/.test(markerDetailSrc));
   assert('showDetailModal captures trigger before opening', /showDetailModal[\s\S]*?rememberModalTrigger\(\)/.test(markerDetailSrc));
   assert('closeModal restores trigger on close', /function closeModal\(\)[\s\S]*?restoreModalTrigger\(\)/.test(markerDetailSrc));
   assert('rememberModalTrigger exported', markerDetailSrc.includes('export function rememberModalTrigger'));
   assert('rememberModalTrigger on window', /window\s*,\s*\{[\s\S]*?rememberModalTrigger/.test(viewsSrc));
-  assert('wearable detail modal captures trigger', wearablesSrc.includes('window.rememberModalTrigger?.()'));
+  assert('wearable detail modal captures trigger', wearablesDetailSrc.includes('window.rememberModalTrigger?.()'));
   assert('restoreModalTrigger guards against detached elements', /document\.contains\(el\)/.test(markerDetailSrc));
 
   // ═══════════════════════════════════════════════
